Sandbox Nilesoft Only solution
In latest version of Windows and Sandbox the Nilesoft only method doesn't work. It results in the same thing as the  hybrid option so the Nilesoft menu only shows when you click "show more options". I altered the command so it works properly now but there is a lag waiting for explorer to restart before it the menu shows like it should.